alert component cutof bug
