+++ webapp/channels/src/components/alert_banner/alert_banner.scss
@@ -3,7 +3,6 @@
 
 .AlertBanner {
     display: flex;
-    overflow: hidden;
     align-items: flex-start;
     padding: 14px;
     border: 1px solid;
